--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -53,7 +53,7 @@ function fill (headers, object) {
   // 1. If object is a sequence, then for each header in object:
   // Note: webidl conversion to array has already been done.
   if (Array.isArray(object)) {
-    for (let i = 0; i < object.length; i++) {
+    for (let i = 0; i < object.length; ++i) {
       const header = object[i]
       // 1. If header does not contain exactly two items, then throw a TypeError.
       if (header.length !== 2) {
@@ -72,9 +72,8 @@ function fill (headers, object) {
     // 2. Otherwise, object is a record, then for each key → value in object,
     //    append (key, value) to headers
     const keys = Object.keys(object)
-    for (let i = 0; i < keys.length; i++) {
-      const key = keys[i]
-      headers.append(key, object[key])
+    for (let i = 0; i < keys.length; ++i) {
+      headers.append(keys[i], object[keys[i]])
     }
   } else {
     throw webidl.errors.conversionFailed({
@@ -440,7 +439,7 @@ class Headers {
     const cookies = this[kHeadersList].cookies
 
     // 3. For each name of names:
-    for (let i = 0; i < names.length; i++) {
+    for (let i = 0; i < names.length; ++i) {
       const [name, value] = names[i]
       // 1. If name is `set-cookie`, then:
       if (name === 'set-cookie') {
@@ -449,7 +448,7 @@ class Headers {
 
         // 2. For each value of values:
         // 1. Append (name, value) to headers.
-        for (let j = 0; j < cookies.length; j++) {
+        for (let j = 0; j < cookies.length; ++j) {
           headers.push([name, cookies[j]])
         }
       } else {