--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -20,15 +20,7 @@ const kHeadersSortedMap = Symbol('headers map sorted')
  * @param {number} code
  */
 function isHTTPWhiteSpaceCharCode (code) {
-  switch (code) {
-    case 0x00a:
-    case 0x00d:
-    case 0x009:
-    case 0x020:
-      return true
-    default:
-      return false
-  }
+  return code === 0x00a || code === 0x00d || code === 0x009 || code === 0x020;
 }
 
 /**