--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -20,7 +20,7 @@ const kHeadersSortedMap = Symbol('headers map sorted')
  * @param {number} code
  */
 function isHTTPWhiteSpaceCharCode (code) {
-  return code === 0x00a || code === 0x00d || code === 0x009 || code === 0x020;
+  return code === 0x00a || code === 0x00d || code === 0x009 || code === 0x020
 }
 
 /**