--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -36,7 +36,7 @@ function headerValueNormalize (potentialValue) {
   while (j > i && isHTTPWhiteSpaceCharCode(potentialValue.charCodeAt(j - 1))) --j
   while (i > j && isHTTPWhiteSpaceCharCode(potentialValue.charCodeAt(i))) ++i
 
-  return i === 0 && j === potentialValue.length ? potentialValue : potentialValue.slice(i, j)
+  return i === 0 && j === potentialValue.length ? potentialValue : potentialValue.substring(i, j)
 }
 
 function fill (headers, object) {