--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -34,7 +34,7 @@ function headerValueNormalize (potentialValue) {
   let i = 0; let j = potentialValue.length
 
   while (j > i && isHTTPWhiteSpaceCharCode(potentialValue.charCodeAt(j - 1))) --j
-  while (i > j && isHTTPWhiteSpaceCharCode(potentialValue.charCodeAt(i))) ++i
+  while (j > i && isHTTPWhiteSpaceCharCode(potentialValue.charCodeAt(i))) ++i
 
   return i === 0 && j === potentialValue.length ? potentialValue : potentialValue.substring(i, j)
 }