--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -114,9 +114,10 @@ function appendHeader (headers, name, value) {
   //    forbidden response-header name, return.
 
   // 7. Append (name, value) to headers’s header list.
+  return headers[kHeadersList].append(name, value)
+
   // 8. If headers’s guard is "request-no-cors", then remove
   //    privileged no-CORS request headers from headers
-  return headers[kHeadersList].append(name, value)
 }
 
 class HeadersList {