--- a/lib/fetch/webidl.js
+++ b/lib/fetch/webidl.js
@@ -399,11 +399,6 @@ webidl.nullableConverter = function (converter) {
 
 // https://webidl.spec.whatwg.org/#es-DOMString
 webidl.converters.DOMString = function (V, opts = {}) {
-  // Note: avoid re-stringify
-  if (typeof V === 'string') {
-    return V
-  }
-
   // 1. If V is null and the conversion is to an IDL type
   //    associated with the [LegacyNullToEmptyString]
   //    extended attribute, then return the DOMString value