<!--
Before submitting a Pull Request, please read our contribution guidelines, which
can be found at CONTRIBUTING.md in the repository root.

For code changes:
1. Include tests for any bug fixes or new features.
2. Update documentation if relevant.
3. Ensure that tests and linting pass.

You will also need to ensure that your contribution complies with the
Developer's Certificate of Origin, outlined in CONTRIBUTING.md
-->

<!-- List the issues this resolves or relates to here (if applicable) -->

## Rationale

This eliminates the `Headers` performance bottleneck.

- `for...of` is slower than `for(;;)`.
- `Object.entries` is slower than  `Object.keys`.
- Reduced the number of variables in the function `ByteString`.
- When trimming `HTTP whitespace`, using RegExp is slow, so change to using `charCodeAt` instead.

<!-- Briefly explain the purpose of this pull request, if not already
justifiable with the above section. If it is, you may omit this section. -->

## Changes

<!-- Write a summary or list of changes here -->


### Features

<!-- List the new features here (if applicable), or write N/A if not -->

N/A

### Bug Fixes

<!-- List the fixed bugs here (if applicable), or write N/A if not -->

N/A

### Breaking Changes and Deprecations

<!-- List the breaking changes (changes that modify the existing API) and
deprecations (removed features) here -->

N/A

## Benchmark


```js
// bench.js
const Headers = require("./lib/fetch/headers.js").Headers;
const benchmark = require("benchmark");

const suite = new benchmark.Suite();

const headers = {
  "Content-Type": "application/json",
  "X-Content-Type": "application/json",
  Date: "Wed, 01 Nov 2023 00:00:00 GMT",
  "X-NodeJS-Version": "v21.1.0",
  "Powered-By": "NodeJS",
  "Content-Encoding": "gzip",
  "Set-Cookie": "__Secure-ID=123; Secure; Domain=example.com",
  "Content-Length": "150",
  Vary: "Accept-Encoding, Accept, X-Requested-With",
};

const headersEntries = Object.entries(headers);


suite.add("new Headers - record", () => {
  new Headers(headers);
});

suite.add("new Headers - entries", () => {
  new Headers(headersEntries);
});

suite.on("cycle", function (event) {
  console.log(String(event.target));
});

// run async
suite.run({ async: true });
```

- main
```
new Headers - record x 144,160 ops/sec ±1.91% (85 runs sampled)
new Headers - entries x 156,706 ops/sec ±1.88% (83 runs sampled)
```

- this patch
```
new Headers - record x 179,333 ops/sec ±4.49% (72 runs sampled)
new Headers - entries x 170,316 ops/sec ±2.39% (80 runs sampled)
```

## Status

<!-- KEY: S = Skipped, x = complete -->


- [x] I have read and agreed to the [Developer's Certificate of Origin][cert]
- [x] Tested
- [ ] Benchmarked (**optional**)
- [ ] Documented
- [x] Review ready
- [ ] In review
- [ ] Merge ready

[cert]: https://github.com/nodejs/undici/blob/main/CONTRIBUTING.md