@metcoder95 2023-11-02
It seems there's a small regression:

**Branch**
![image](https://github.com/nodejs/undici/assets/10200776/9cb0ab27-e9ba-4f38-81de-d8f7bb99f605)


**Main**
![image](https://github.com/nodejs/undici/assets/10200776/ba8a925b-a1eb-4292-9979-d3d02bca6f6a)

Running once more to verify