@tsctx 2023-11-03
> How much does performance improve by only removing the regex?
> 
> I don't really like the idea of removing the for..of loops because I don't find it maintainable, in the sense that no one will block a PR that uses it over the standard for loop. I don't mind the change and I'm not asking you to revert it though, just mentioning.

This is just a benchmark for removing RegExp.

```
new Headers - record x 161,831 ops/sec ±2.69% (83 runs sampled)
new Headers - entries x 168,051 ops/sec ±4.50% (80 runs sampled)
```