@KhafraDev 2023-11-03
so it looks like a majority of the improvement is from removing the regex and removing Object.entries?