@tsctx 2023-11-04
Hello!
New improvements have been made.
The conversion to `ByteString` seems to have already been completed by the webidl converter `sequence<sequence<ByteString>> or record<ByteString, ByteString>`, but it is being converted again.
See the new benchmark below.
[`9d838c5`](https://github.com/tsctx/undici/tree/4592203e7b71721de1d46a8921e00b9687a5c092)
```
new Headers - record x 185,007 ops/sec ±2.16% (82 runs sampled)
new Headers - entries x 165,584 ops/sec ±1.75% (83 runs sampled)
```
[`f0aaa32`](https://github.com/tsctx/undici/tree/f0aaa32ca5925b1ad7d56f4c3bc09685ab34f232)
```
new Headers - record x 255,866 ops/sec ±2.29% (80 runs sampled)
new Headers - entries x 224,652 ops/sec ±1.87% (88 runs sampled)
```